feat: Enhance viewport data handling and field vector requests
- Introduced new logic to manage field frame envelopes and latest field frames in session runtime.
- Added scaling functionality for field components in viewport data bridge.
- Updated viewport bridge activity to ensure magnetic texture shader data remains active regardless of selected quantity.
- Implemented queuing mechanism for field vector requests to handle concurrent requests more efficiently.
- Adjusted resource cache size to optimize memory usage.
- Added tests for new functionality in viewport field roles and preview selection state.
- Refactored vector rendering logic to support Float32 arrays and scaling.

+++ apps/web/components/preview/r3f/FdmInstances.tsx
@@ -39,7 +39,7 @@ export interface ResolvedFdmGridBounds {
 
 interface FdmInstancesProps {
   grid: [number, number, number];
-  vectors: Float64Array | null;
+  vectors: Float32Array | null;
   geometryMode: boolean;
   activeMask: boolean[] | null;
   settings: {
